[expected.un.object] Remove unnecessary subclause nesting
All the other class templates in <expected> are at the rSec3 level, but
std::unexpected is below a mostly useless rSec3 [expected.unexpected].
This subclause has two children, [expected.un.general] which is a single
sentence, and [expected.un.object] which defines the class template and
its members. If we merge the single sentence from [expected.un.general]
into the same subclause as the class synopsis then we can get remove the
unnecessary nesting. As a nice side effect, this also gets rid of
"object" in the [expected.un.object] stable name, which doesn't really
make sense in context.