Add examples in [format.string.escaped]
The s2 examples is reproduced by using the cyrillic latex
package.

The s6 example cannot be sanely reproduced with the
latex engine used.

Instead, use N + combining tilde as an example.
To make sure the LHS does not appear to be related to
normalization, spell out the combining grapheme extend
codepoint.