[diff] Uniform style for examples
For C++ differences, examples are always preceded by the phrase "For
example:", do not use the usual [Example n: ... -- end example] style,
and always appear in the "Effects on original feature" paragraph.

For differences with C, a different set of styles is used (examples
being part of paragraphs such as "Change" and "Rationale"), and that
subclause remains unchanged by this commit.