[format.string.std] Reorder std-format-spec field descriptions. (#5246)
Moves the wording describing the zero-padding before the description of
the width; matching the order of the fields in the std-format-spec.

The original order was introduced in the initial <format> paper P0645R10 "Text Formatting".
Since both fields had one paragraph of description, it wasn't too noticeable. P1868R2 "🦄 width:
clarifying units of width and precision in std::format" expanded the wording of the width.
Now it's not so easy to find the description of the zero-padding field.