[mdspan.layout.{left,right}.overview] Replace "see below" with condition
The condition is spelled out in the item descriptions already, and the
class synopses seem to simply have been inconsistent with that in the
incoming paper P0009R18. Since the "see below"s are never referenced
explicitly, we just replace them with the actual conditions, which is
also how the surrounding members are presented.