[basic.life]/8.3 don't put const in the middle of "complete object" words of power
"Complete object" is a defined term (https://eel.is/c++draft/basic#def:complete_object). Putting const in the middle of that makes it unclear if we are referring to a const-qualified "complete object" or a complete const object with some other meaning of "complete" such as "with a complete type". Using the order "const complete object" avoids that ambiguity. It also matches the word order used in https://eel.is/c++draft/basic.life#10 to make it UB to create a new object on top of a "const complete object", unless it has dynamic storage duration.