Update references to Unicode Standard & Annexes.
Currently we have a normative reference to ISO 10646 (Unicode 13.0),
a floating normative reference to UAX44 (= Unicode 15.0),
bibligraphy items for Unicode (14.0),
text segmentation UAX (12.0), and identifier syntax (13.0).

This align the bibligraphy items and mentions of Unicode to refer
to 15.0.

Mention of Unicode 14 in the specification of print are also updated
to refer to Unicode 15.

No further changes are necessary to comply to the updated standard
and annexes.