[utility.arg.requirements] turn list of identifiers into a bullet list
Lists providing the key for names used throughout a subclause are difficult
to follow when buried as a single sentence in a paragraph.  This PR updates
the key to tables 29-36 in [utility.arg.requirements] to a bullet list
following the examples set elsewhere.

Each key is now consistently introduced with the word "denotes", following
the precedent set by similar bullet lists.

It is not at all coincidental that this also resolves pagination issues
at the end of the page, so that the floating tables to not float into the
middle of a sentance in the following section.