Add/fix references to func.require, func.invoke
from [meta] and [threads]

In particular, not knowing where to find the definition of *INVOKE* (it's in the library index, but not the main index) is highly frustrating when reading is_invocable et al.

Corrected references in [threads] and added some more for completeness.

There's a bunch of uses of 'invoke' in [algorithm] and [ranges] that could also have references added, but that feels less necessary.