[container.adaptors] uses_allocator specializations consistently use "Alloc".
This is not intended to trigger (nor un-trigger) any blanket wording;
it's purely cosmetic, aligning the parameter names for the flat_foo
containers' `uses_allocator` specializations, with the parameter
names for stack/queue/priority_queue.

Intuitively, `Alloc` is the more appropriate name because this type
is not necessarily something that qualifies as an Allocator; it is
the same kind of type as the `Alloc` constructor parameters in
[queue.defn] for example.

My understanding is that the old wording's choice to use `Alloc` in
one specialization and `Allocator` in another is _not_ motivated by
any intentional difference in semantics between the two specializations;
it's just a bit of inconsistency that we can patch up editorially.