[coro.generator.promise] Use "equivalent to `throw`, ..."
Our style guide says that "equivalent to <expression>" is the
appropriate style for an expression of type `void`, which is the case
for `throw`. This avoids the awkward situation of having "equivalent
to: `throws;`" appear in the middle of a paragraph without proper
terminal punctuation.

Fixes NB US 119 (C++23 CD).