[utilities] Consistently index all member typedefs in the utilites header
Audited the file `utilities.tex` for indexing of member typedefs.  Several
idioms were in use, and not all typedef names were indexed.  Adopted what
seemed to be the preferred approach of using the `@\libmember{}@` style of
definition within the class definition itself, so that omissions become
obvious if more typedef names are added in the future.

Please provide feedback on whether indexing the explicit template
specialization for <void> types is the correct approach.  It pulls out
the name that is not present for the primary template, and then the
ranges library adds some non-template classes with the same name.

Noted that `char_type` for `basic_format_arg` is a private member marked
as exposition only, so should really be `char-type` with the expos-id
font.  After checking with the original paper, will supply a separate
follow-up ticket for that --- this comment is only to confirm that the
omission from the index in this patch is intentional.