[depr.conversions.string] remove redundent respecification (#6004)
There are several typedef names defined precisely in the deprecated code conversion
facets classes that are redundantly respecified in text. Nowhere else in the library
does this; any specified typedef names are _see below_ definitions, not repeats of
the class definition. Needless redundancy is always a risk of divergence, however
small, so remove the respecification in text form.