[basic.life] Move definition of "before" and "after" from bottom to top of subclause
The last paragraph of this subclause changes the definition of English words
used throughout the preceding paragraphs.  While it might be preferable
to replace all such usage with the new definitions, that would be a Core issue,
see paragraph 6 for an example of awkward usage.  Hence, we move the
redefinition to the start of the subclause so we know how to read this text
from the start.