[re.results.general] Add type of `match_results::const_iterator` to index of imp def
The type of most container iterators is records in the index of implementation
defined types, but somehow the `match_results` iterator in the <regex> clause
was missed.

After applying this patch, I grepped for any other occurrences of "\\impdef}"
and the only remaining (and intentional) match is the definition of this
macro in macros.tex, so there should be no more latent cases to find.