[span.iterators] Fix cross-reference to container iterators
The current cross-reference is to [containers.requirements], which
is the whole containers requirements clause, including not just
general containers, but also allocator-aware, reversible, sequence,
associative, and unodered associative containers.  It seems very
unlikely that the cross-reference expects to be the intersection
of all of those.

Rather, the reference seems to intend just the [containers.reqmts]
subclause, which adds just two useful constraints: random access
iterators should support the 3-way comparison operator, and the
interoperabiity of container iterators and const_iterators.

As a drive-by, fixed the \indexlibrarymember macro by ordering
the type, span, before the member, iterator, as is the convention.
This has no impact on the index, as entries in both directions are
created.  However, it makes the wording of index macros across the
clause consistent.