[container.reqmts] Relocate more container requirements
As part of rewriting the container requirements tables as text,
several paragraphs of general container requirements were buried
in the new subsection for more specific requirements such as
reversible containers or allocator-aware containers.  This seems
to have happened when general container requirements followed one
of the container requirements tables that are noe expressed as
text.

This PR carefully restores text to the general container requirements
clause, in the original order they appeared in C++20, and audits that
there are no further paragraphs to move that do not have outstanding
pull requests.