[container.reqmts] Move requirements to the right place (#6199)
As part of the replacement of the container requirements tables with text,
several paragraphs of general container requirements were buried
in the new subsections for more specific requirements such as
reversible containers or allocator-aware containers.

This seems to have happened when general container requirements
followed one of the container requirements tables that are now expressed
as text. This change carefully restores text to the general container requirements
clause, in the original order they appeared in C++20.