[depr.atomics.volatile] Use tcode to call out template name
There is no such thing as an "atomic specialization", so the
only valid reading of this text is that "atomic" refers to
the class template `atomic`.

This PR makes the intent clear with code font and a cross
reference, and I believe is editorial.  However, the LWG
chair might want to confirm whether the change of font
counts as a normative change that deserves LWG review.