[algorithms.parallel.defns] Insert new paragraph number for example
The example currently in paragraph 3 has its own list, whole numbering
extends that of the normative list preceding it.  This can be
confusing at first sight, with a one-line defintion apparently
inserted into the middle of the list.  By adding a new pnum for
the example, the list numbering in the example restarts at 4.1,
reducing confusion.