Remove section 2 from [sequence.reqmts]
I have been requested to make this a direct request rather than putting it in P0447.

Despite the cute poem, section 2 should be removed because the note is misleading and over-simplified. See Appendix J of P0447 for a more complete basic guide to container selection. Such a guide is too large to fit in the standard and will be subject to changes in architecture in terms of usefulness. The section conflates time complexity and performance, and oddly recommends vector as a default without mentioning its time complexity, unlike the other containers. Time complexity's relevance to both performance and latency is subject to architectural differences. If the committee chose to re-include such a note, it should simply describe time complexity as below (being removed may be better as it creates less confusion):

[Note 1: The sequence containers offer the programmer different complexity trade-offs. vector has amortized constant time complexity for insertions at the end of the sequence, and constant time complexity for erasures at the end of the sequence. hive, list and forward_list have constant time complexity for insertions and erasures at any point in the sequence, however for hive the user cannot specify the insertion position. deque has constant time complexity for insertions and erasures at the beginning or end of the sequence. -end note]