[alg.ends.with]: drop_view should be views::drop
We don't want programmers doing CTAD on `ranges::drop_view(args...)`,
so we shouldn't put it in the as-if-by code either. What we expect
people to do (and what we intend vendors to do here) is `views::drop`.

Tim Song points out that when `r1` is a move-only view such as an
`owning_view`, then `views::drop(r1, N)` alone is ill-formed because
`views::all(r1)` is ill-formed. We have to manually wrap the argument
`r1` in a `ranges::ref_view` (for which there is no user-facing name)
in order to handle `owning_view` correctly.