[temp.func.order] Explicitly specify to only add extra first argument if we don't have an explicit object argument
In [temp.func.order] paragraph 3, it is written that we need to add a new first parameter for all member functions for the purposes of template deduction. However, it only makes sense if we don't have an explicit object parameter.

If we have an explicit object parameter, we shouldn't add a new (implicit) first parameter because our first parameter is explicitly written.