[memory.syn][specialized.algorithms] Prefer trailing returns on complex signatures
This proposed change follows the observation that when function
template signatures get complicated, especially with long names
for return types, and non-trivial requires clauses, it can be
very hard to spot the function name and return type among the
sea of tokens, in order to start mentally parsing the function.

Here we propose using the trailing return type form of function
declaration in those cases, and apply it consistently to the
special algorithms in the '<memory>' header as a non-trivial
sample size to gauge interest in a more complete PR that would
cover the whole ranges library.