[memory.syn] Move specification macros to appropriate subclause
The specification macros _POINTER_OF_ and _POINTER_OF_OR_ do not appear
in the synopsis of the '<memory>' header, but are in fact used only to
specify behavior of the smart pointer adapters.  Created a new
Definitions heading in that subclause, and moved the text there.

In addition, add the two specification macros to the main index, and
also the index of library names, following the example set by _INVOKE_.
Given the choice, I would rather all of these entries were in just the
main index, but following precedent.