[implimits] Split lines to match new ordering
This change may be semi-normative.  It splits lines that defined
several quantities into separate lines that can then be correctly
ordered.  The secondary benefit is that these quantities may now
be updated independantly.  The new risk is that these quantities
that were previously guaranteed to be the same can now vary in a
future standard.