[lex.charset] Extract universal-character-name grammar to new subclause
The grammar for universal-character-name is oddly sandwiched into the
middle of the subcluase talking about the different character sets used
by the standard.  To improve the flow, extract that grammar into its own
subclause.

In the extraction, I make two other clarifying changes.  First, describe
this new subclause as 'a way to name any element of the of the tranlation
character set using just the basic character set' rather than simply
'a way to name other characters'.  Secondly, remove the 'one of' in the
grammar where there is only one option to choose.