[std] Reorganize the initial Core clauses
This PR colocates [lex], [cpp], and modules to put all
the parts that talk about assembling and translating a
program together.  In doing so, it rearranges the
subclauses in [lex] and introduces subclauses that can
be cross-references for each phase of translation.

Metadata is introduced to identify the first and last
core clause when cross-references want the first/last
property rather than the specific clause itself.

The subclause on comments, [lex.comment], is merged
into the new [lex.phase.3] as comments feature only
during translation phase three.