[lex.phases] replace term 'input file' with 'source file' in phase 1
The only part of the standard that refers to input files is
phase 1 of the phases of translation.  Everywhere else, we
use the term source file, that has a clear definition in
the preceding [lex.separate] clarifying that we do not
necessarily mean text files in a hosted filing system.