[lex.separate][module.unit] move definitions of program and translation unit
The definition of program at the top of [basic.link] should move to
the front of [lex.separate] so that it is defined before its first
usage, and also clarifies that is what the phases of translation produce.

Similarly, move the definition of the grammar production translation-unit
to the top of the only clause whose grammar actually uses it, [module.unit].

Finally, retitle [basic.link] as just Linkage, rather than
Programs and linkage.