[lex] Reorganize contents to follow grammar and phases of translation
This PR purely moves existing words around, and does not create any
new content.  It would be the precursor to a larger change set that
might integrate [cpp] into lex, or move it adjactent, and similarly
move [modules] adjactent.

This may be more significant than we want to land in C++26, but I
offer it while time is still available, and as inspiration for the
C++29 reorg otherwise.