[cpp.predefined] Sort prefined macros
This change also puts the text requiring the feature macros to
be predefined adjacent to the table of feature macros, at the
end of the list.  It also moves the '__STDCPP_THREADS' macro
from the conditionally defined list into the main list, as it
is specified in exactly the same form as the five macros for
extended floating point types in the main list.