[text.encoding.aliases] Add note about what isn't required
Make it explicit that `copyable` is intended, rather than `semiregular`,
and that the return types of `begin()` and `end()` can differ.

Also remove a FIXME comment for something that doesn't need fixing.
These requirements on a type are specified in our usual form, it would
be wrong to use _Remarks_: or _Requires_: here.