[basic.types.general] Clarify confusing phrase "object and value representation"
There is no such thing as the "object and value representation" of a type;
there is only the "object representation" and the "value representation."
Refactor the confusing sentence accordingly. This seems to make one of the
two new sentences potentially nonsensical, but no worse than it was when it
was both grammatically confusing *and* potentially nonsensical.