[ios.init] Remove unused Init::init_cnt static member
The text that made use of this variable was removed by LWG1123 and has
not been present in the WP since N3090. The effects of Init construction
and destruction are specified entirely without the use of this variable,
so it serves no purpose now.

LWG1123 even noted that this variable could be dropped, stating "Some
editorial changes are expected (in addition to the proposed wording) to
remove init_cnt from Init." Let's do that. Better late than never.