[lex] Prepare to merge phase 4 preprocessor into lex
This commit makes substantial changes to the organization of
the Core clauses, without changing wording.  Notably, [lex]
is now titled as Program construction rather than Lexical
conventions, as its contents are significatly more than mere
conventions, and with the phases of translation detail how a
program is constructed and built.

[lex] itself is now more clearly modeled after the phases of
translation for each of the level 2 subheadings, grouping the
relevant specification for how to process each phases that
needs more than the simple text in the phases themsevles.

The whole Core sections are reorganized, promoting overload
and exception handling above classes and templates.  The idea
is that overloading is relevant to understanding how functions
are called, while functions are well specified before we get
to classes and templates. Exception handling moves forward as
the last control flow, which is relevant for functions before
we get to defining classes.  Classes still precede templates,
as clearly we support both function and class templates.

For neatness, we integrate the fix for tautonyms in stable
labels. Finally, we try to disentagle some of the wording
around basic definitions of names etc.