[lex][cpp] Merge [cpp] into [lex]
This commit completes the planned Core reorganization by
merging the preprocessor properly into Program construction,
placed consistently for phase 4 of translation. The subclauses
of [cpp] are similarly reorganized to put the more primitive
functionality first, and putting the module related text up
front, consistent with the order of the grammar.  We finally
bring the prohibition on defining keywords as macros out of
the library introduction and directly into the specification
for keywords.

This is around the limit of what can be achieved before
rewriting text to further reduce duplication, and to make
the distinction of preprocessor-identifiers and TU-identifiers
clear.