[except.uncaught] Tidy the specification for uncaught exceptions
Several concurrent fixes.  First include the normative wording
that 'uncaught_exceptions' returns the number of uncaught
exceptions *on the current thread*.  This wording is present
in the core language.

Then move the core wording for when an exception is uncaught
directly into the text that talks about caught and uncaught
exceptions.  In the process, turn the reference to
 into a note, so that there is only
one normative specification.

Finally, remove [except.uncaught] as it is now empty, and add
the missing descriptive sentence to the intro paragraph of
[except.special.general].  This happens to produce quite a
pleasing page-break, but that is just luck, not design.