[except.terminate] Better describe the  function
While 'std:terminate' was originally conceived as the way to
report failures in the exception handling machinery, it has
evolved to become a more general tool for reporting unrecoverable
failures in the C++ runtime.  This rewording attempts to address
that evolving design, and in doing so addresses the outstanding
%FIXME% that the current text is not adequately descriptive in
the first place.