[except.handle] group all paragraphs on searching for handler
This commit moves all of the paragraphs involved in the search for a
handler for an exception into a single logical sequence.

After this change, [except.spec] deals only with specifying the
'noexcept' function decorator and its interaction with the
'noexcept' operator, and contains no text regarding exceptions
themselves.  It might be appropriate to move that subclause into
the [dcl] structure at a future date.