[lex.charset] Introduce parent subclause [lex.char] for character sets and UCNs (#7067)
The grammar for universal-character-name is oddly sandwiched into the
middle of the subclause talking about the different character sets used
by the standard.  To improve the flow, extract that grammar into its own
subclause.

In the extraction, I make three other clarifying changes.  First, describe
this new subclause as 'a way to name any element of the of the translation
character set using just the basic character set' rather than simply
'a way to name other characters'. Then, merge the sentence on where universal
characters are prohibited into the new intro sentence describing universal
characters, to make clear that there is no contradiction between nominating
a character, and how that character can be used. Finally, remove the 'one of'
in the grammar where there is only one option to choose.