[lex.header] Modernize text around header names
The footnote better belongs in the main text as a regular note.
To make the notes flow consistently, switch the order of the
note and normative text in the first paragraph to lead with the
normative text.  Finally, take this opportunity to clarify the
specific characters called out using the Unicode nomenclature
used throughout this clause.