[allocator.requirements.general] Remove redundant template syntax
In the Cpp17Allocator requirements, all identifiers name specific
types, not type-dependent code in a template, even if they are
template instantiations.  Hence, all use of `::template` is redundant.
Not in the sense of we are awaiting LWG to review P2150 for policy,
but in the sense these have been redundant since C++11, and were
outright ill-formed prior to that.