[lex.separate] Redistribute second comment
The second comment in [lex.separate] is both too specific, and not specific
enough.  Move the last sentence about use of separately translated TUs to
the end of the first note, where it seems most appropriate.  Move the rest
of the note to the end of translation phase 8, where we are in a position
to talk about preserving the separately translated TUs *and instantiation units*
for linking in phase 9.