[expr.await] Define 'suspend' as a term of power for coroutines
Many parts of the standard refer to whether a coroutine is suspended as if
it were a term of power, but there is no \defn that defines it as such.
This PR places the definition on the only use that I could find where a
suspended state was not defined in terms of other suspended states, so
may be the transition into suspension.  It is likely that folks with a
better eye to Core wording will know a better place, that may need a little
rewording to sneak in the word 'suspended' to properly define it.