[lex.pptoken] Separate out keyword placeholders
The _preprocessing-token_ term is defined as one of another set of
grammar productions, with the three placeholder tokens standing out
as specific single tokens.

Moving the placeholder tokens into their own production simplifies
both the _preprocessing-token_ and _keyword_ grammar productions,
plus it guarantees a consistency for the set of placeholder between
both places.  It also makes it easier for [cpp] to directly
cross-reference where needed.

Pedantically, the term placeholder token itself is problematic as
tokens are a phase 7 concern, where they become keywords.  Within
the earlier phases where these placeholders are used the correct
term is a preprocessing token, so strictly these should be
placeholder preprocessing tokens, which is a mouthful.  So renaming
the term to placeholder keyword, which is hopefully less of a
concern as there are no keywords in phase 3--6 to confuse with
phase 7 keywords.

Finally, clean up the informative list in text of the various kinds
of preprocessing tokens, rendering as a list in the source (one kind
per line) and adding a cross-reference to each such item.